e_commerce/requirements.txt
Django==5.2.4
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
django-redis==5.4.0
channels==4.1.0
channels-redis==4.2.0
daphne==4.1.2
asgiref==3.8.1
python-dotenv==1.0.1